split modules for Tabs; improve tests
